clean up and move code
